applications/menubar-videoconf-app/requirements.txt
rumps
requests
pyperclip
python-decouple
